pyproject.toml
[tool.poetry]
name = "pmdb"
version = "0.1.0"
description = ""
authors = ["Tristan Kernan <tristanmkernan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-allauth = "^0.51.0"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"
Pillow = "^9.2.0"
django-guardian = "^2.4.0"
django-taggit = "^3.0.0"
python-magic = "^0.4.27"
django-htmx = "^1.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
